добавляет вебпак, делает разбивку на модули
