+++ src/view/show-more.js
@@ -1,0 +1,7 @@
+const createShowMoreTemplate = () => {
+  return (
+    `<button class="films-list__show-more">Show more</button>`
+  );
+};
+
+export {createShowMoreTemplate};
